Move components from file to package
