+++ invenio_rdm_records/services/components/__init__.py
@@ -1,0 +1,21 @@
+# -*- coding: utf-8 -*-
+#
+# Copyright (C) 2021 CERN.
+#
+# Invenio-RDM-Records is free software; you can redistribute it and/or modify
+# it under the terms of the MIT License; see LICENSE file for more details.
+
+
+"""High-level API for working with RDM service components."""
+
+from .access import AccessComponent
+from .metadata import MetadataComponent
+from .parent import ParentRecordAccessComponent
+from .pids import ExternalPIDsComponent
+
+__all__ = (
+    'AccessComponent',
+    'ExternalPIDsComponent',
+    'MetadataComponent',
+    'ParentRecordAccessComponent',
+)
